refactor(backend): change methods name to comply with RAD
