+++ src/main/java/it/astromark/chat/service/TicketServiceImpl.java
@@ -66,7 +66,7 @@ public class TicketServiceImpl implements TicketService {
     @Override
     @Transactional
     @PreAuthorize("hasRole('TEACHER') || hasRole('PARENT') || hasRole('SECRETARY')")
-    public List<MessageResponse> getMessages(@NotNull Ticket ticket) {
+    public List<MessageResponse> getMessageList(@NotNull Ticket ticket) {
 
         var messages = messageRepository.findByTicket(ticket);
         if (!messages.isEmpty())
